pyproject.toml
[project]
name = "game"
version = "0.1.0"
description = ""
authors = [{ name = "DenzelPenzel", email = "leopolnow@gmail.com" }]
license = { text = "MIT" }
readme = "README.md"
keywords = ["python", "game"]
dependencies = [
    "pytest==8.3.3",
    "ruamel-yaml==0.16.10",
    "pre-commit==2.18.1"
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 120
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
line_length = 120
multi_line_output = 3
include_trailing_comma = true
use_parentheses = true
ensure_newline_before_comments = true
combine_as_imports = true
filter_files = true
skip = ["setup.py"]
